+++ lib/firefile.dart
@@ -1,2 +1,1 @@
 export 'src/firefile.dart';
-export 'src/models/firefile_task.dart';